CONFIG-1987: move load_from_env to query and storage config
